refactor: modularize app
