+++ index.html
@@ -120,6 +120,6 @@
     DEFAULT_RADIUS_METERS: 5000
   };
 </script>
-<script src="app.js?v=3.9"></script>
+<script type="module" src="app.js?v=3.9"></script>
 </body>
 </html>
